requirement.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
black==22.6.0
bleach==5.0.1
click==8.1.3
Django==4.0.6
django-markdownx==3.0.1
django-phonenumber-field==6.3.0
django-summernote==0.8.20.0
django-tagging==0.5.0
importlib-metadata==4.12.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.51
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.8.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.1
webencodings==0.5.1
zipp==3.8.1
